refactor: 전체적인 폴더구조 수정 및 style 관련 import 규칙 수정
