+++ src/pages/login/components/login-logo/index.tsx
@@ -1,0 +1,8 @@
+import * as S from './styles';
+import LoginTitle from '@/assets/login/Kakao_Login_Title.svg';
+
+const LoginLogo = () => {
+  return <S.LoginLogo src={LoginTitle} alt="카카오 CI" />;
+};
+
+export default LoginLogo;
