+++ src/pages/login/components/login-logo/styles.ts
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled';
+
+export const LoginLogo = styled.img`
+  width: 88px;
+  color: rgb(51, 51, 51);
+`;
