+++ src/pages/main/components/select-gift-type/index.tsx
@@ -1,5 +1,5 @@
 import { Grid } from '@/components/common/layouts/Grid';
-import * as Styles from '../styles';
+import * as S from './styles';
 import { TypeData } from '@/constants/GiftData';
 import { Container } from '@/components/common/layouts/Container';
 import { useNavigate } from 'react-router-dom';
@@ -7,18 +7,18 @@ import { useNavigate } from 'react-router-dom';
 const SelectGiftType = () => {
   const navigate = useNavigate();
   return (
-    <Styles.SelectGiftLayout>
+    <S.SelectGiftLayout>
       <Container>
         <Grid columns={6}>
           {TypeData.map((data, index) => (
-            <Styles.GiftTypeContainer onClick={() => navigate(`/theme/${data.type}`)} key={index}>
-              <Styles.GiftImg src={data.image} alt={data.title} />
-              <Styles.GiftTypeTitle>{data.title}</Styles.GiftTypeTitle>
-            </Styles.GiftTypeContainer>
+            <S.GiftTypeContainer onClick={() => navigate(`/theme/${data.type}`)} key={index}>
+              <S.GiftImg src={data.image} alt={data.title} />
+              <S.GiftTypeTitle>{data.title}</S.GiftTypeTitle>
+            </S.GiftTypeContainer>
           ))}
         </Grid>
       </Container>
-    </Styles.SelectGiftLayout>
+    </S.SelectGiftLayout>
   );
 };
