+++ src/pages/my-account/MyAccountPage.tsx
@@ -1,22 +1,11 @@
-import { useNavigate } from 'react-router-dom';
-import * as Styles from './styles';
-import { PATHS } from '@/constants/Path';
+import * as S from './styles';
+import MyInfo from './components/my-info';
 
 const MyAccountPage = () => {
-  const userName = sessionStorage.getItem('authToken') || 'User';
-  const navigate = useNavigate();
-
-  const HandlerLogout = () => {
-    sessionStorage.removeItem('authToken');
-    navigate(PATHS.MAIN);
-  };
-
   return (
-    <Styles.MyAccountPageLayout>
-      {userName}님 안녕하세요!
-      <Styles.BoxSpacer />
-      <Styles.LogoutButton onClick={HandlerLogout}>로그아웃</Styles.LogoutButton>
-    </Styles.MyAccountPageLayout>
+    <S.MyAccountPageLayout>
+      <MyInfo />
+    </S.MyAccountPageLayout>
   );
 };
